+++ src/test/java/nextstep/auth/AuthorizationTokenExtractorTest.java
@@ -1,5 +1,6 @@
 package nextstep.auth;
 
+import nextstep.auth.util.AuthorizationTokenExtractor;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
